+++ Data/Cil.Todo.Data/Domain/State.cs
@@ -7,7 +7,7 @@ namespace Cil.Todo.Data.Domain
         public string Name { get; set; }
         public int CountryId { get; set; }
         public Country Country { get; set; }
-        public int DisplayOrder { get; set; }
+        public short DisplayOrder { get; set; }
 
         public ICollection<City> Cities { get; set; }
     }
